requirements.txt
absl-py==2.2.1
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.2
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
hpsklearn==0.1.0
hyperopt==0.2.7
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.1
kiwisolver==1.4.8
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
nose==1.3.7
numba==0.61.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pmlb==1.0.1.post3
protobuf==5.29.4
py4j==0.10.9.9
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.2
shap==0.47.1
six==1.17.0
slicer==0.0.8
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.0
threadpoolctl==3.6.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==3.0.0
scipy==1.15.2